relation to categories added, product tests added

+++ src/main/java/dev/mark/jewelsstorebackend/products/ProductDTO.java
@@ -4,5 +4,5 @@ public class ProductDTO {
     public String productName;
     public String description;
     public String image;
-    public int price;
+    public Long price;
 }
